+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This change
 ## [1.0.0-BETA16-SNAPSHOT] 
 - added vrksasna draft for session management
 - a very minor extension of clojisr internals to support that
+- removed GCReference (and tech.resource dependency) - replaced by `proxy` object.
 
 ## [1.0.0-BETA15] - 2020-07-12
 - [hot fix] datasets are not transfered as `data.frame` due to implementation of `map` interfaces in t.m.ds 3.01